+++ spikeinterface/widgets/legacy_mpl_widgets/__init__.py
@@ -19,7 +19,7 @@ from .drift import (plot_drift_over_time, DriftOverTimeWidget,
                     plot_displacement, DisplacementWidget)
 
 # waveform/PC related
-from .unitwaveforms import plot_unit_waveforms, plot_unit_templates
+# from .unitwaveforms import plot_unit_waveforms, plot_unit_templates
 from .unitwaveformdensitymap import plot_unit_waveform_density_map, UnitWaveformDensityMapWidget
 from .amplitudes import plot_amplitudes_timeseries, plot_amplitudes_distribution
 from .principalcomponent import plot_principal_component
