+++ spikeinterface/widgets/legacy_mpl_widgets/amplitudes.py
@@ -3,7 +3,7 @@ from matplotlib import pyplot as plt
 
 from .basewidget import BaseWidget
 
-from ..toolkit import compute_spike_amplitudes
+from ...toolkit import compute_spike_amplitudes
 from .utils import get_unit_colors
