+++ spikeinterface/widgets/legacy_mpl_widgets/principalcomponent.py
@@ -2,7 +2,7 @@ import numpy as np
 from matplotlib import pyplot as plt
 
 from .basewidget import BaseWidget
-from ..toolkit import WaveformPrincipalComponent, compute_principal_components
+from ...toolkit import WaveformPrincipalComponent, compute_principal_components
 
 
 class PrincipalComponentWidget(BaseWidget):
